refactor: refatora estrutura de pastas
