+++ jedi/.gitignore
@@ -34,4 +34,4 @@ build/
 
 /docker
 
-
+.idea/
