+++ jedi/src/main/java/com/ifce/jedi/config/StartupInitializer.java
@@ -3,8 +3,8 @@ package com.ifce.jedi.config;
 import com.ifce.jedi.dto.Banner.BannerDto;
 import com.ifce.jedi.dto.Banner.BannerItemUrlDto;
 import com.ifce.jedi.dto.Header.HeaderDto;
-import com.ifce.jedi.model.UserRole;
-import com.ifce.jedi.model.User;
+import com.ifce.jedi.model.User.UserRole;
+import com.ifce.jedi.model.User.User;
 import com.ifce.jedi.repository.UserRepository;
 import com.ifce.jedi.service.BannerService;
 import com.ifce.jedi.service.HeaderService;
