+++ jedi/src/main/java/com/ifce/jedi/dto/Authenticator/AuthenticatorDto.java
@@ -1,4 +1,4 @@
-package com.ifce.jedi.dto;
+package com.ifce.jedi.dto.Authenticator;
 
 public record AuthenticatorDto(String email, String password) {
 }
