+++ jedi/src/main/java/com/ifce/jedi/dto/Authenticator/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.ifce.jedi.dto;
+package com.ifce.jedi.dto.Authenticator;
 
 public record LoginResponseDto(String token) {
 }
