+++ jedi/src/main/java/com/ifce/jedi/dto/Authenticator/RegisterDto.java
@@ -1,6 +1,6 @@
-package com.ifce.jedi.dto;
+package com.ifce.jedi.dto.Authenticator;
 
-import com.ifce.jedi.model.UserRole;
+import com.ifce.jedi.model.User.UserRole;
 
 public record RegisterDto(String email, String password, UserRole role) {
 }
