+++ jedi/src/main/java/com/ifce/jedi/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.ifce.jedi.repository;
 
-import com.ifce.jedi.model.User;
+import com.ifce.jedi.model.User.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.core.userdetails.UserDetails;
